refactor(web): simplify per-device provider settings
Cleanup pass over the providers panel from parallel reuse, simplification,
efficiency, and altitude reviews:

- Share the connection-phase dot/ping mapping from ConnectionStatusDot instead
  of a third inline copy
- Extract the React Compiler hook harness and element-tree walker into
  src/test and reuse them from both new test files
- Move interval/background-activity helpers and PolicyTooltip out of the
  SettingsPanels component file into SettingsPanels.logic and settingsLayout
- Drop the setState-in-useEffect selection sync; the effective selection is
  derived, so a device that reconnects regains its prior selection
- Only mount the primary-session atom when a browser session is actually
  gated by it; remote devices skip the SWR fetch entirely
- Guard runProviderUpdate re-entry with a ref instead of a state-updater flag
- Treat a failed session fetch as transport trouble rather than denied access
- Reuse EnvironmentConnectionPhase, isElectron, and the shared
  EMPTY_SERVER_PROVIDERS constant; remove the indefinite loading spinner
